Allow DataResult to lazily evaluate error message
This commit is a first-step, automatic IDE refactors incoming!

A lot of time in datafixing time is spent creating error messages. These often can include stringified NBT/json which is fairly expensive to build.

These errors are most often never used, where the error is only used to determine when to fall back to another path.