Add parameter to buildOptimized() to filter which types need prebuilt fixers
Most of the fixers (roughly half) are optimised for types that are never used. Many types exist to represent an element contained within another type. For example, entities are contained within chunks: but an entity never needs to be updated alone. The chunk fixer function inlines the recursive functions that apply to entities, so the function that applies only to entities is never required.