Add comment to highlight bug with intermediate types construction when sorting nested optics
Temporary comment to point it out for now, will be resolved along with the following commit.

The construction of the intermediate type assumes that the optic is a single level, however the rule allows it to be nested. These swaps are technically still correct, as they operate on different data, but the intermediate type is, in this case, incorrectly produced.